+++ script/bcld_vendor.sh
@@ -97,8 +97,11 @@ function update_cert () {
 ## To set NSSDB
 function bcld_set_nssdb () {
 	list_item "Configuring certificate database for: ${BCLD_VENDOR^^}"
-	/usr/bin/sudo /usr/bin/mkdir -p "${NSSDB}"
-	/usr/bin/sudo /usr/bin/cp ${HOME}/nssdb/${BCLD_VENDOR}/* "${NSSDB}"
+	list_entry
+	/usr/bin/sudo /usr/bin/mkdir -pv "${NSSDB}"
+	/usr/bin/sudo /usr/bin/cp -v ${HOME}/nssdb/${BCLD_VENDOR}/{cert9.db,key4.db,pkcs11.txt} "${NSSDB}"
+	/usr/bin/sudo /usr/bin/chown -Rv "${USER}:${USER}" "${NSSDB}"
+	list_catch
 }
 
 ## To get certificates